+++ src/plugins/storage-lokijs/rx-storage-instance-loki.ts
@@ -23,7 +23,6 @@ import type {
     RxStorageBulkWriteResponse,
     RxStorageQueryResult,
     RxJsonSchema,
-    MangoQuery,
     LokiStorageInternals,
     RxStorageInstanceCreationParams,
     LokiDatabaseSettings,
@@ -167,7 +166,6 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
         }
 
         const ret: RxStorageBulkWriteResponse<RxDocType> = {
-            success: [],
             error: []
         };
 
@@ -196,7 +194,6 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
 
         categorized.bulkInsertDocs.forEach(writeRow => {
             localState.collection.insert(flatClone(writeRow.document));
-            ret.success.push(writeRow.document);
         });
         categorized.bulkUpdateDocs.forEach(writeRow => {
             const docId = writeRow.document[this.primaryPath];
@@ -209,7 +206,6 @@ export class RxStorageInstanceLoki<RxDocType> implements RxStorageInstance<
                 }
             );
             localState.collection.update(writeDoc);
-            ret.success.push(writeRow.document);
         });
         localState.databaseState.saveQueue.addWrite();